Dataset 2 {NAM}	R Documentation

Multi-environmental trial

Description

Data collected from the SoyNAM population in Indiana 2013-2015.

Usage

  data(met)

Details

Data provided by Rainey Lab for Soybean Breeding and Genetics, Purdue University. Genotypic matrix (Gen) have 4240 SNPs. The data frame Obs contains the soybean id (ID), environment (Year), field location (Block,Row,Col) and three phenotypes: grain yield (YLD), days to maturity (DTM) and average canopy closure (ACC).

Author(s)

Alencar Xavier, Ben Hall and Katy Rainey

References

Xavier, A., Hall, B., Hearst, A.A., Cherkauer, K.A. and Rainey, K.M., 2017. Genetic Architecture of Phenomic-Enabled Canopy Coverage in Glycine max. Genetics, 206(2), pp.1081-1089.

[Package NAM version 1.8.0 Index]