SNP H2 {NAM}	R Documentation

SNP heritability

Description

Calculates the ability of markers to carry a gene (Foneris et al. 2015). The index is also an indicator of Mendelian segregation.

Usage

snpH2(gen,K=NULL)

Arguments

gen
Numeric matrix containing the genotypic data. A matrix with n rows of observations and (m) columns of molecular markers.

K
Optional. Numeric matrix containing the genetic relationship matrix. A square and symmetric matrix (nxn) with the same observations as the matrix gen, also in the same order. If not provided, the kinship matrix is estimated from the genotypic matrix.

Value

Numeric vector containing the heritability of each markers. Foneris et al. (2015) recommends to avoid using markers with index lower than 0.98.

Author(s)

Alencar Xavier

References

Forneris, N. S., Legarra, A., Vitezica, Z. G., Tsuruta, S., Aguilar, I., Misztal, I., & Cantet, R. J. (2015). Quality Control of Genotypes Using Heritability Estimates of Gene Content at the Marker. Genetics 199(3):675-681.

Examples

## Not run: 
data(tpod)
Heritability=snpH2(gen)
plot(Heritability,chr=chr)

## End(Not run)

[Package NAM version 1.8.0 Index]