Dataset 1 {NAM}	R Documentation

Tetra-seed Pods

Description

Two soybean bi-parental crosses phenotyped for the percentage of pods containing four seeds

Usage

  data(tpod)

Details

Soybean nested association panel with 2 families (fam) containing 196 indiviuals. Genotypic matrix (gen) have 376 SNP across 20 chromosome (chr). Phenotypic information (y) regards the proportion of four seed pods. Data provided by Rainey Lab for Soybean Breeding and Genetics, Purdue University.

Author(s)

Alencar Xavier and Katy Rainey

[Package NAM version 1.8.0 Index]